[refactor] extract useTheme as a custom hook file
